Added MP40 placeholder weapon
Created new files to add a weapon in ETQW. The GDF soldier is equipped
with it. The weapon has a script, a GUI icon, GUI strings (fr, en),
sounds, and it's own ammo type. Currently it lacks the 3D model because
I've lost the normalmap and also because the model needs changes. To be
updated!